refactor: more indicative file naming for range finder
